+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed issues which caused `Camera`, `PostProcessingSprite` and `PickingHitArea` to not work correctly when renderer resolution was more than 1.
 - Fixed an issue which caused `screenToRay` on `Camera` to return incorrect results when `orthographic = true`
+- Fixed orientation for default IBL lookup texture.
 
 ## [0.9.8] - 2021-09-23
 ### Added